backend/pyproject.toml
[project]
name = "ebay-movie-assist"
version = "0.1.0"
authors = [
    { name = "Ryan Davis", email = "ryancraigdavis@gmail.com" },
]
description = "ebay movie assistant backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "boto3>=1.34.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "aiofiles>=23.2.0",
    "pillow>=10.1.0",
    "httpx>=0.25.0",
    "pandas>=2.1.0",
]

[build-system]
requires = ["hatchling"]

build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["src/ebay_movie_assist"]

[tool.hatch.metadata]
allow-direct-references = true
